# expression:implement vectorized evaluation for builtinInstrBinarySig