<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->
PCP #12103
### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for builtinInstrBinarySig

### What is changed and how it works?
```
BenchmarkVectorizedBuiltinStringFunc/builtinInstrBinarySig-VecBuiltinFunc-8         	   76005	     17651 ns/op	       0 B/op	       0 allocs/op    

BenchmarkVectorizedBuiltinStringFunc/builtinInstrBinarySig-NonVecBuiltinFunc-8      	   18814	     67997 ns/op	       0 B/op	       0 allocs/op     

BenchmarkVectorizedBuiltinStringFunc/builtinInstrBinarySig-VecBuiltinFunc#01-8      	   44893	     26276 ns/op	       0 B/op	       0 allocs/op    

BenchmarkVectorizedBuiltinStringFunc/builtinInstrBinarySig-NonVecBuiltinFunc#01-8   	   17289	     71267 ns/op	       0 B/op	       0 allocs/op  
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test